Fixed bugs in reading hwp format and added test example

+++ HwpFile/HwpDoc/OLEdoc/DirectoryEntry.h
@@ -13,13 +13,13 @@ class CDirectoryEntry
 	int m_nLeftSiblingID;
 	int m_nRightSiblingID;
 	int m_nChildID;
-	long m_lClsID1;
-	long m_lClsID2;
+	long long m_lClsID1;
+	long long m_lClsID2;
 	int m_nStateBit;
-	long m_lCreationTime;
-	long m_lModifiedTime;
+	long long m_lCreationTime;
+	long long m_lModifiedTime;
 	int m_nStartingSectorID;
-	long m_lStreamSize;
+	long long m_lStreamSize;
 	VECTOR<int> m_arSecNums;
 public:
 	CDirectoryEntry();
@@ -30,19 +30,19 @@ public:
 	void SetLeftSiblingID(int nLeftSiblingID);
 	void SetRightSiblingID(int nRightSiblingID);
 	void SetChildID(int nChildID);
-	void SetClsID1(const long& lClsID1);
-	void SetClsID2(const long& lClsID2);
+	void SetClsID1(const long long& lClsID1);
+	void SetClsID2(const long long& lClsID2);
 	void SetStateBit(int nStateBit);
-	void SetCreationTime(const long& lCreationTime);
-	void SetModifiedTime(const long& lModifiedTime);
+	void SetCreationTime(const long long& lCreationTime);
+	void SetModifiedTime(const long long& lModifiedTime);
 	void SetStartingSectorID(int nStartingSectorID);
-	void SetStreamSize(const long& lStreamSize);
+	void SetStreamSize(const long long& lStreamSize);
 
 	int GetLeftSiblingID() const;
 	int GetRightSiblingID() const;
 	int GetChildID() const;
 	int GetStartingSectorID() const;
-	long GetStreamSize() const;
+	long long GetStreamSize() const;
 	VECTOR<int> GetSecNums() const;
 
 	VECTOR<int>& GetSecNums();
